feat : clinet가 주식 전체 조회 및 가격에 변동이 있을 시 정보를 받아볼 수 있도록 기능 구현

+++ src/main/java/com/example/stocksimulation/domain/entity/Stock.java
@@ -14,10 +14,10 @@ import lombok.NoArgsConstructor;
 public class Stock {
     @Id
     private String code;
-    private int price;
+    private long price;
     private String name;
 
-    public void updatePrice(int price) {
+    public void updatePrice(long price) {
         this.price = price;
     }
